# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: EthixAI API
  description: |
    Ethical AI Governance Platform API - Real-time bias detection, SHAP explanations,
    and comprehensive monitoring for financial decision-making systems.
  version: '1.0.0'
  contact:
    name: EthixAI Support
    email: support@ethixai.com
servers:
  - url: http://localhost:5000
    description: Local development
  - url: https://api.ethixai.com
    description: Production
...
paths:
...
  /api/tenants:
    get:
      tags: [Auth]
      summary: List tenants (admin only)
      operationId: listTenants
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of tenants
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Tenant'

    post:
      tags: [Auth]
      summary: Create tenant (admin only)
      operationId: createTenant
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                settings:
                  type: object
      responses:
        '201':
          description: Tenant created
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Tenant:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        settings:
          type: object
        createdAt:
          type: string
          format: date-time

    # ==================== COMMON SCHEMAS ====================
...